Added contact, and new emoji
in the discount scrolling thingy i added a blue money sign with wings i think. for the contact i just did a picture of some people and a google map picture of somewhere random. Also in the places button something happened while i was trying to move it to the bottom. by the way i think we kinda finished but you can add more things if you want like for decoration but the main things that we needed are done.